Parameterize buf compatibility target
This commit updates the buf compatibility check so it targets latest
release (v0.12.9), as it's the one we target compatibility with.

Also, parameterize it so we can arbitrary target other versions, if we
are doing some manual checking, specially in the beginning when we may
want to validate compatibilities for skip release upgrades.

Lastly, introduce `checkbuf` target so we can run buf linter without the
rest.